Move tests to Core from Interpreter
