+++ src/tests/Microsoft.PowerFx.Interpreter.Tests/RecalcEngineTests.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -343,25 +342,6 @@ namespace Microsoft.PowerFx.Tests
 
             Assert.False(result.IsSuccess);
             Assert.StartsWith("Error 4-4: Expected an operand", result.Errors.First().ToString());
-        }
-
-        [Fact]
-        public void CheckParseErrorCommaSeparatedLocale()
-        {
-            var engine = new RecalcEngine(new PowerFxConfig(CultureInfo.GetCultureInfo("it-IT")));
-            var result = engine.Parse("3.145");
-
-            Assert.False(result.IsSuccess);
-            Assert.StartsWith("Error 2-5: Unexpected character", result.Errors.First().ToString());
-        }
-
-        [Fact]
-        public void CheckParseSuccessCommaSeparatedLocale()
-        {
-            var engine = new RecalcEngine(new PowerFxConfig(CultureInfo.GetCultureInfo("de-DE")));
-            var result = engine.Parse("Function(args; separated; by; semicolons) + 123,456");
-
-            Assert.True(result.IsSuccess);
         }
 
         [Fact]
